lwe_lattice/requirements.txt
numpy==2.2.6
pyjwt==2.10.1
psutil==7.0.0
matplotlib==3.9.2